+++ src/au/org/paperminer/dao/IGenericDAO.java
@@ -72,7 +72,7 @@ public interface IGenericDAO<T, ID extends Serializable> {
 	 * 
 	 * @return the number of entities
 	 */
-	int countAll();
+	Long countAll();
 
 	/**
 	 * Count entities based on an example.
